acquisition_program/requirements.txt
opencv-python==4.10.0.84
numpy==1.24.4
pyserial==3.5